+++ mobile/android/android-components/docs/changelog.md
@@ -16,6 +16,10 @@ permalink: /changelog/
   * Added `endOfMenuAlwaysVisible` property/parameter to `BrowserMenuBuilder` constructor and to `BrowserMenu.show` function.
     When is set to true makes sure the bottom of the menu is always visible, this allows use cases like [#3211](https://github.com/mozilla-mobile/android-components/issues/3211).
 
+* **browser-search**
+  * Added `getProvidedDefaultSearchEngine` to `SearchEngineManager` to return the provided default search engine or the first
+    search engine if the default is not set. This allows use cases like [#3344](https://github.com/mozilla-mobile/android-components/issues/3344).
+
 * 🛑 Removed deprecated components (See blog posting):
   * feature-session-bundling
   * ui-progress